+++ app/src/main/java/app/revanced/integrations/music/patches/components/PlayerFlyoutMenuFilter.java
@@ -11,7 +11,8 @@ public final class PlayerFlyoutMenuFilter extends Filter {
         addIdentifierCallbacks(
                 new StringFilterGroup(
                         Settings.HIDE_FLYOUT_MENU_3_COLUMN_COMPONENT,
-                        "music_highlight_menu_item_carousel.eml"
+                        "music_highlight_menu_item_carousel.eml",
+                        "tile_button_carousel.eml"
                 )
         );
     }