+++ app/src/main/java/app/revanced/integrations/music/patches/player/PlayerPatch.java
@@ -13,8 +13,9 @@ import java.util.Arrays;
 
 import app.revanced.integrations.music.settings.Settings;
 import app.revanced.integrations.music.shared.VideoType;
+import app.revanced.integrations.music.utils.VideoUtils;
 
-@SuppressWarnings({"unused", "ResultOfMethodCallIgnored"})
+@SuppressWarnings({"unused"})
 public class PlayerPatch {
     private static final int MUSIC_VIDEO_GREY_BACKGROUND_COLOR = -12566464;
     private static final int MUSIC_VIDEO_ORIGINAL_BACKGROUND_COLOR = -16579837;
@@ -145,14 +146,16 @@ public class PlayerPatch {
         return Settings.HIDE_FULLSCREEN_SHARE_BUTTON.get() ? 0 : original;
     }
 
-    public static void setShuffleState(int buttonState) {
+    public static void setShuffleState(Enum<?> shuffleState) {
         if (!Settings.REMEMBER_SHUFFLE_SATE.get())
             return;
-        Settings.SHUFFLE_SATE.save(buttonState);
+        Settings.ALWAYS_SHUFFLE.save(shuffleState.ordinal() == 1);
     }
 
-    public static int getShuffleState() {
-        return Settings.SHUFFLE_SATE.get();
+    public static void shuffleTracks() {
+        if (!Settings.ALWAYS_SHUFFLE.get())
+            return;
+        VideoUtils.shuffleTracks();
     }
 
     public static boolean rememberRepeatState(boolean original) {
@@ -163,11 +166,15 @@ public class PlayerPatch {
         return Settings.REMEMBER_SHUFFLE_SATE.get();
     }
 
+    public static boolean restoreOldCommentsPopUpPanels() {
+        return restoreOldCommentsPopUpPanels(true);
+    }
+
     public static boolean restoreOldCommentsPopUpPanels(boolean original) {
         if (!Settings.SETTINGS_INITIALIZED.get()) {
             return original;
         }
-        return !Settings.RESTORE_OLD_COMMENTS_POPUP_PANELS.get();
+        return !Settings.RESTORE_OLD_COMMENTS_POPUP_PANELS.get() && original;
     }
 
     public static boolean restoreOldPlayerBackground(boolean original) {