+++ app/src/main/java/app/revanced/integrations/music/patches/video/CustomPlaybackSpeedPatch.java
@@ -49,6 +49,7 @@ public class CustomPlaybackSpeedPatch {
 
     private static void resetCustomSpeeds(@NonNull String toastMessage) {
         Utils.showToastLong(toastMessage);
+        Utils.showToastShort(str("revanced_extended_reset_to_default_toast"));
         Settings.CUSTOM_PLAYBACK_SPEEDS.resetToDefault();
     }