+++ app/src/main/java/app/revanced/integrations/youtube/patches/components/ShortsButtonFilter.java
@@ -84,6 +84,11 @@ public final class ShortsButtonFilter extends Filter {
                 "shorts_info_panel_overview"
         );
 
+        StringFilterGroup stickers = new StringFilterGroup(
+                Settings.HIDE_SHORTS_STICKERS,
+                "stickers_layer.eml"
+        );
+
         StringFilterGroup liveHeader = new StringFilterGroup(
                 Settings.HIDE_SHORTS_LIVE_HEADER,
                 "immersive_live_header"
@@ -122,7 +127,7 @@ public final class ShortsButtonFilter extends Filter {
         addPathCallbacks(
                 suggestedAction, actionBar, joinButton, subscribeButton, metaPanelButton,
                 paidPromotionButton, pausedOverlayButtons, channelBar, fullVideoLinkLabel,
-                videoTitle, reelSoundMetadata, infoPanel, liveHeader
+                videoTitle, reelSoundMetadata, infoPanel, liveHeader, stickers
         );
 
         //
@@ -158,6 +163,10 @@ public final class ShortsButtonFilter extends Filter {
                 new ByteArrayFilterGroup(
                         Settings.HIDE_SHORTS_REMIX_BUTTON,
                         "reel_remix_button"
+                ),
+                new ByteArrayFilterGroup(
+                        Settings.DISABLE_SHORTS_LIKE_BUTTON_FOUNTAIN_ANIMATION,
+                        "shorts_like_fountain"
                 )
         );
 
@@ -194,8 +203,9 @@ public final class ShortsButtonFilter extends Filter {
                         "yt_outline_location_point_"
                 ),
                 new ByteArrayFilterGroup(
-                        Settings.HIDE_SHORTS_SAVE_SOUND_BUTTON,
-                        "yt_outline_list_add_"
+                        Settings.HIDE_SHORTS_SAVE_MUSIC_BUTTON,
+                        "yt_outline_list_add_",
+                        "yt_outline_bookmark_"
                 ),
                 new ByteArrayFilterGroup(
                         Settings.HIDE_SHORTS_SEARCH_SUGGESTIONS_BUTTON,
@@ -209,8 +219,19 @@ public final class ShortsButtonFilter extends Filter {
                         Settings.HIDE_SHORTS_USE_TEMPLATE_BUTTON,
                         "yt_outline_template_add"
                 ),
+                new ByteArrayFilterGroup(
+                        Settings.HIDE_SHORTS_GREEN_SCREEN_BUTTON,
+                        "shorts_green_screen"
+                ),
                 useThisSoundButton
         );
+    }
+
+    private boolean isEverySuggestedActionFilterEnabled() {
+        for (ByteArrayFilterGroup group : suggestedActionsGroupList)
+            if (!group.isEnabled()) return false;
+
+        return true;
     }
 
     @Override
@@ -247,6 +268,9 @@ public final class ShortsButtonFilter extends Filter {
         }
 
         if (matchedGroup == suggestedAction) {
+            if (isEverySuggestedActionFilterEnabled()) {
+                return super.isFiltered(path, identifier, allValue, protobufBufferArray, matchedGroup, contentType, contentIndex);
+            }
             // Suggested actions can be at the start or in the middle of a path.
             if (suggestedActionsGroupList.check(protobufBufferArray).isFiltered()) {
                 return super.isFiltered(path, identifier, allValue, protobufBufferArray, matchedGroup, contentType, contentIndex);