+++ app/src/main/java/app/revanced/integrations/youtube/patches/utils/PlayerTypeHookPatch.java
@@ -1,8 +1,11 @@
 package app.revanced.integrations.youtube.patches.utils;
 
+import android.view.View;
+
 import androidx.annotation.Nullable;
 
 import app.revanced.integrations.youtube.shared.PlayerType;
+import app.revanced.integrations.youtube.shared.ShortsPlayerState;
 import app.revanced.integrations.youtube.shared.VideoState;
 
 @SuppressWarnings("unused")
@@ -23,6 +26,27 @@ public class PlayerTypeHookPatch {
         if (youTubeVideoState == null) return;
 
         VideoState.setFromString(youTubeVideoState.name());
+    }
+
+    /**
+     * Injection point.
+     * <p>
+     * Add a listener to the shorts player overlay View.
+     * Triggered when a shorts player is attached or detached to Windows.
+     *
+     * @param view shorts player overlay (R.id.reel_watch_player).
+     */
+    public static void onShortsCreate(View view) {
+        view.addOnAttachStateChangeListener(new View.OnAttachStateChangeListener() {
+            @Override
+            public void onViewAttachedToWindow(@Nullable View v) {
+                ShortsPlayerState.set(ShortsPlayerState.OPEN);
+            }
+            @Override
+            public void onViewDetachedFromWindow(@Nullable View v) {
+                ShortsPlayerState.set(ShortsPlayerState.CLOSED);
+            }
+        });
     }
 }